+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - add `merged_person_from_login` method to BackendApiConnector
+- RestrictedTextLanguage allowing only EN or DE if confidence >=0.75
 
 ### Changes
+- detect_language does not support FR, RU or ES anymore.
 - bumped cookiecutter template to https://github.com/robert-koch-institut/mex-template/commit/a67c71
 
 ### Deprecated